// corinth-code-frontend/src/app/models/swapi/swapit-starships.ts
export interface SwapiStarship {
  MGLT: number;
  cargo_capacity: number;
  consumables: string;
  cost_in_credits: number;
  created: Date;
  crew: number;
  edited: Date;
  films: string[];
  hyperdrive_rating: number;
  length: number;
  manufacturer: string;
  max_atmosphering_speed: number;
  model: string;
  name: string;
  passengers: number;
  pilots: string[];
  starship_class: string;
  url: string;
}